Minor changes in common files. Printf doesn't work :(. Uninitialized value somewhere. Check valgrind
Took 38 seconds

+++ libft/t_vec.h
@@ -73,7 +73,7 @@ int					ft_string_fit(t_string **str);
 int					ft_string_push_back(t_string **str, char c);
 int					ft_string_push_back_s(t_string **str_ptr, char *str);
 int					ft_string_push_back_n_c(t_string **str_ptr,
-		size_t n, char c);
+		long n, char c);
 
 int					ft_string_set_value(t_string **str, size_t n, char filler,
 				char value);
